refactor(rd-parser): split parser.rs into focused submodules
- mod.rs: Parser struct, errors, core section parsing, token utilities
- macro_dispatch.rs: \macro dispatcher and macro name parsing
- arg_parsers.rs: bracketed/text/inline argument helpers
- node_parsers.rs: individual macro parsers (list, tabular, link, etc.)
- tests.rs: test suite extracted from the main file
